Refactor Privacy Filter and remove unused exports
